Refactored MainPage.js (peer review needed)
- created a custom hook for the dragging functionality (src/customHooks/useDrag.js)

- refactored logic in a weird location (into src/feature/utilities/utilitiesSlice.js)
